% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task}
\alias{task}
\title{The combination of the material to be worked and the tool to work it with}
\usage{
task(name, tool, material)
}
\arguments{
\item{name}{The name of the task}

\item{tool}{The function used in the task}

\item{material}{The object being worked by the tool}
}
\description{
The combination of the material to be worked and the tool to work it with
}
